backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
pdf2docx==0.5.6
docx2pdf==0.1.8